# Refactor nonbonded energy