--- a/docs/_docs/energy.md
+++ b/docs/_docs/energy.md
@@ -113,23 +113,6 @@ combinations:
       protein water: 60
 ~~~
 
-### OpenMP Control
-
-If compiled with OpenMP, the following keywords can be used to control parallelisation
-for non-bonded interactions. The best combination depends on the simulated system size and
-composition. Currently, parallelisation is disabled by default.
-
-~~~ yaml
-- nonbonded:
-    openmp: [g2g, i2all]
-~~~
-
-`openmp`  | Description
---------- | -------------------------------------------
-`g2g`     | Distribute on a molecule-to-molecule basis 
-`i2all`   | Parallelise single particle energy evaluations
-
-
 ## Electrostatics
 
  `coulomb`             |  Description