--- a/src/atomdata.h
+++ b/src/atomdata.h
@@ -74,9 +74,32 @@ void from_json(const json &j, std::vector<AtomData> &v);
 
 extern std::vector<AtomData> atoms; //!< Global instance of atom list
 
+/**
+ * @brief Find the first element with a member `name` matching the input.
+ * @see obtainName()
+ * @param rng a range
+ * @param name a name to look for
+ * @return an iterator to the first element or `last` if not found
+ */
 template <class Trange> auto findName(Trange &rng, const std::string &name) {
     return std::find_if(rng.begin(), rng.end(), [&name](auto &i) { return i.name == name; });
-} //!< Returns iterator to first element with member `name` matching input
+}
+
+/**
+ * @brief Find the first element with a member `name` matching the input. Throw an exception if not found.
+ * @see findName()
+ * @throw std::range_error if such an element not found
+ * @param rng a range
+ * @param name a name to look for
+ * @return an iterator to the first element
+ */
+template <class Trange> auto obtainName(Trange &rng, const std::string &name) {
+    const auto result = findName(rng, name);
+    if(result == rng.end()) {
+        throw std::out_of_range(name + " not found");
+    }
+    return result;
+}
 
 template <class Trange> std::vector<int> names2ids(Trange &rng, const std::vector<std::string> &names) {
     std::vector<AtomData::Tid> index;