# Description

- Separate the logic determining which particles have changed upon a move from a general particle pairing logic. Introduce a basic pairing policy. Another pairing policies , e.g., OMP optimized, may be added later to allow faster nonbonded energy computation.
- Extensively document the code.

As unintended side effect, some example tests exhibit a rather huge speed-up (GCC 9.2 on Linux).
| example  | master (s) | pull request (s) |
| ----------- | ----------- |  ----------- |
| `swapconf` |  5.54 |  1.95 |
| `polymers` | 17.97 |  9.68 |
| `speciation_group` | 70.92 |  8.95 |
| `water-ewald-NOCHECKS` | 53.78 | 11.45 |
| `cluster` | 43.53 | 23.70 |

Note that OpenMP in particle pairing is effectively disabled with this pull request until OpenMP pairing strategy is introduced.

## Checklist

- [x] `make test` passes with no errors
- [x] the source code is well documented
- [ ] new functionality includes unittests
- [x] the user-manual has been updated to reflect the changes

Ready for _Rebase and merge_ strategy.