@rc83 2020-02-25
I have updated the table in the description to contain also Clang besides GCC. The table contains the major differences, either between master and PR, or GCC and Clang.