@rc83 2020-02-26
NonbondedCache tested using an example provided by Giulio.